refactor(contributions): extract ContributionsList component
